+++ package.json
@@ -55,6 +55,7 @@
     "@typescript-eslint/parser": "^5.18.0",
     "babel-loader": "^8.2.4",
     "bootstrap": "^4.0.0",
+    "concurrently": "^7.6.0",
     "cross-env": "^7.0.3",
     "eslint": "^8.12.0",
     "eslint-config-prettier": "^8.5.0",
@@ -64,6 +65,7 @@
     "gh-pages": "^3.2.3",
     "jest": "^27.5.1",
     "microbundle": "^0.14.2",
+    "nodemon": "^2.0.20",
     "prettier": "^2.6.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
