+++ src/lib/DataTableFilterRow/DataTableFilterRow.tsx
@@ -1,9 +1,8 @@
 import React from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faSearch, faTimes } from "@fortawesome/free-solid-svg-icons";
 import { Input } from "reactstrap";
-import { DataTableColumnDescription, DataTableActions, Filters } from "../DataTable/DataTableInterfaces";
+import { DataTableColumnDescription, DataTableActions, Filters, FilterTranslations } from "../DataTable/DataTableInterfaces";
 import { ColumnFilterType, ActionsPosition } from "../DataTable/DataTableTypes";
+import { ActionsHeaderFilterCell } from "../DataTable/Actions/ActionsHeaderFilterCell";
 
 interface DataTableFilterRowProps<T> {
   columns: DataTableColumnDescription<T>[];
@@ -16,62 +15,22 @@ interface DataTableFilterRowProps<T> {
   onSearch(): void;
 }
 
-interface FilterTranslations {
-  searchToolTip: string;
-  clearSearchToolTip: string;
-}
-
 export function DataTableFilterRow<T>({
   columns,
   actions,
-  actionsPosition = ActionsPosition.Left,
+  actionsPosition,
   filterPossible = true,
   translations,
   getFilterRefs,
   setFilterRef,
   onSearch,
 }: DataTableFilterRowProps<T>) {
-  const Actions = () => (
-    <>
-      {actions && (
-        <th className={actions.className} style={actions.style}>
-          <FontAwesomeIcon
-            style={{ cursor: "pointer", marginBottom: "4px", marginRight: "5px" }}
-            title={translations.searchToolTip}
-            icon={faSearch}
-            onClick={() => {
-              onSearch();
-            }}
-          />
-          <FontAwesomeIcon
-            style={{ cursor: "pointer", marginBottom: "4px", marginRight: "5px" }}
-            title={translations.clearSearchToolTip}
-            icon={faTimes}
-            onClick={() => {
-              const filterRefs = getFilterRefs();
-              if (filterRefs) {
-                Object.values(filterRefs).forEach((ref) => {
-                  if (ref instanceof HTMLSelectElement) {
-                    // eslint-disable-next-line no-param-reassign
-                    ref.value = null as unknown as string;
-                  } else {
-                    // eslint-disable-next-line no-param-reassign
-                    ref.value = "";
-                  }
-                });
-                onSearch();
-              }
-            }}
-          />
-        </th>
-      )}
-    </>
-  );
-
   if (!filterPossible || columns.filter((column) => column.filter).length <= 0) return <React.Fragment />;
   return (
     <tr>
-      {actionsPosition == ActionsPosition.Left && <Actions />}
+      {actionsPosition == ActionsPosition.Left && (
+        <ActionsHeaderFilterCell<T> onSearch={onSearch} translations={translations} actions={actions} getFilterRefs={getFilterRefs} />
+      )}
       {columns.map((column) => (
         <th key={column.dataField}>
           {column.filter && column.filter.filterType === ColumnFilterType.String && (
@@ -103,7 +62,9 @@ export function DataTableFilterRow<T>({
           )}
         </th>
       ))}
-      {actionsPosition == ActionsPosition.Right && <Actions />}
+      {actionsPosition == ActionsPosition.Right && (
+        <ActionsHeaderFilterCell<T> onSearch={onSearch} translations={translations} actions={actions} getFilterRefs={getFilterRefs} />
+      )}
     </tr>
   );
 }
